added the ability to select the teebox when teams are playing off different tees

+++ GolfProto04/View models/Games/TeamTeeBoxViewModel.swift
@@ -39,7 +39,7 @@ struct TeamTeeBoxViewModel: Hashable {
     var teamTeeBoxId: NSManagedObjectID {
         return teamTeeBox.objectID
     }
-    var courseRating: Int16 {
+    var courseRating: Double {
         return teamTeeBox.courseRating
     }
